# modules/apps/data-engine/data-engine-rest-impl/rest-openapi.yaml
components:
    schemas:
...
        DataLayout:
            properties:
                contentType:
                    type: string
                dataDefinitionId:
                    format: int64
                    type: integer
                dataLayoutKey:
                    type: string
                dataLayoutPages:
                    items:
                        $ref: "#/components/schemas/DataLayoutPage"
                    type:
                        array
                dataRules:
                    items:
                        $ref: "#/components/schemas/DataRule"
                    type:
                        array
                dateCreated:
                    format: date-time
                    type: string
                dateModified:
                    format: date-time
                    type: string
                description:
                    additionalProperties:
                        type: object
                    type: object
                id:
                    format: int64
                    type: integer
                name:
                    additionalProperties:
                        type: object
                    type: object
                paginationMode:
                    type: string
                siteId:
                    format: int64
                    type: integer
                userId:
                    format: int64
                    type: integer
            type: object
        DataLayoutColumn:
            properties:
                columnSize:
                    type: integer
                fieldNames:
                    items:
                        type: string
                    type:
                        array
            type: object
        DataLayoutPage:
            properties:
                dataLayoutRows:
                    items:
                        $ref: "#/components/schemas/DataLayoutRow"
                    type:
                        array
                description:
                    additionalProperties:
                        type: object
                    type: object
                title:
                    additionalProperties:
                        type: object
                    type: object
            type: object
        DataLayoutRenderingContext:
            properties:
                containerId:
                    type: string
                dataRecordValues:
                    additionalProperties:
                        type: object
                    type: object
                namespace:
                    type: string
                pathThemeImages:
                    type: string
                readOnly:
                    type: boolean
            type: object
        DataLayoutRow:
            properties:
                dataLayoutColumns:
                    items:
                        $ref: "#/components/schemas/DataLayoutColumn"
                    type:
                        array
            type: object
...
        DataRule:
            properties:
                actions:
                    items:
                        additionalProperties:
                            type: object
                        type: object
                    type:
                        array
                conditions:
                    items:
                        additionalProperties:
                            type: object
                        type: object
                    type:
                        array
                logicalOperator:
                    type: string
                name:
                    additionalProperties:
                        type: object
                    type: object
            type: object
info:
    description:
        "A Java client JAR is available for use with the group ID 'com.liferay', artifact ID
        'com.liferay.data.engine.rest.client', and version '2.0.19'."
    license:
        name: "Apache 2.0"
        url: "http://www.apache.org/licenses/LICENSE-2.0.html"
    title: "Data Engine"
    version: v2.0
openapi: 3.0.1
paths:
...
    "/data-definitions/{dataDefinitionId}/data-layouts":
        delete:
            operationId: deleteDataLayoutsDataDefinition
            parameters:
                - in: path
                  name: dataDefinitionId
                  required: true
                  schema:
                      format: int64
                      type: integer
            responses:
                204:
                    content:
                        application/json: {}
                        application/xml: {}
                    description:
                        ""
            tags: ["DataLayout"]
        get:
            operationId: getDataDefinitionDataLayoutsPage
            parameters:
                - in: path
                  name: dataDefinitionId
                  required: true
                  schema:
                      format: int64
                      type: integer
                - in: query
                  name: keywords
                  schema:
                      type: string
                - in: query
                  name: page
                  schema:
                      type: integer
                - in: query
                  name: pageSize
                  schema:
                      type: integer
                - in: query
                  name: sort
                  required: false
                  schema:
                      type: string
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                items:
                                    $ref: "#/components/schemas/DataLayout"
                                type: array
                        application/xml:
                            schema:
                                items:
                                    $ref: "#/components/schemas/DataLayout"
                                type: array
                    description:
                        ""
            tags: ["DataLayout"]
        post:
            operationId: postDataDefinitionDataLayout
            parameters:
                - in: path
                  name: dataDefinitionId
                  required: true
                  schema:
                      format: int64
                      type: integer
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/DataLayout"
                    application/xml:
                        schema:
                            $ref: "#/components/schemas/DataLayout"
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/DataLayout"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/DataLayout"
                    description:
                        ""
            tags: ["DataLayout"]
...
    "/data-layouts/{dataLayoutId}":
        delete:
            operationId: deleteDataLayout
            parameters:
                - in: path
                  name: dataLayoutId
                  required: true
                  schema:
                      format: int64
                      type: integer
            responses:
                204:
                    content:
                        application/json: {}
                        application/xml: {}
                    description:
                        ""
            tags: ["DataLayout"]
        get:
            operationId: getDataLayout
            parameters:
                - in: path
                  name: dataLayoutId
                  required: true
                  schema:
                      format: int64
                      type: integer
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/DataLayout"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/DataLayout"
                    description:
                        ""
            tags: ["DataLayout"]
        put:
            operationId: putDataLayout
            parameters:
                - in: path
                  name: dataLayoutId
                  required: true
                  schema:
                      format: int64
                      type: integer
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/DataLayout"
                    application/xml:
                        schema:
                            $ref: "#/components/schemas/DataLayout"
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/DataLayout"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/DataLayout"
                    description:
                        ""
            tags: ["DataLayout"]
    "/data-layouts/{dataLayoutId}/context":
        post:
            operationId: postDataLayoutContext
            parameters:
                - in: path
                  name: dataLayoutId
                  required: true
                  schema:
                      format: int64
                      type: integer
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/DataLayoutRenderingContext"
                    application/xml:
                        schema:
                            $ref: "#/components/schemas/DataLayoutRenderingContext"
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                additionalProperties:
                                    type: object
                                type: object
                        application/xml:
                            schema:
                                additionalProperties:
                                    type: object
                                type: object
                    description:
                        ""
            tags: ["DataLayout"]
...
    "/sites/{siteId}/data-layouts/by-content-type/{contentType}/by-data-layout-key/{dataLayoutKey}":
        get:
            operationId: getSiteDataLayoutByContentTypeByDataLayoutKey
            parameters:
                - in: path
                  name: siteId
                  required: true
                  schema:
                      format: int64
                      type: integer
                - in: path
                  name: contentType
                  required: true
                  schema:
                      type: string
                - in: path
                  name: dataLayoutKey
                  required: true
                  schema:
                      type: string
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/DataLayout"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/DataLayout"
                    description:
                        ""
            tags: ["DataLayout"]
...